Juno (#7652)
* Added community layout juno and junonum; removed poorly maintained variants of juno; added junonum-a47.

* Added community layout juno and junonum; removed poorly maintained variants of juno; added junonum-a47.

* updated layout type for juno

* Performing requested changes; renaming a47; fine-tuning of codes in community layout keymaps.

* Fine tuning to requested changes.

* Further tuning to suggestion.

* Update layouts/community/ortho_4x12/juno/keymap.c


* Update layouts/community/ortho_4x12/juno/keymap.c


* Update keyboards/maartenwut/atom47/keymaps/junonum_a47/keymap.c


* Update keyboards/maartenwut/atom47/keymaps/junonum_a47/keymap.c
