+++ keyboards/maartenwut/atom47/keymaps/junonum_a47/readme.md
@@ -1,0 +1,1 @@
+Juno's junonum keymap adapted for Atom47.
