--- keyboards/planck/keymaps/juno/readme.md
@@ -1,8 +1,0 @@
-# Juno Layout
-created by Juno Nguyen
-juno.ngx@gmail.com
-
-## Feature:
-* Pok3r-inspired functional layer.
-* Rearranged numerics for maximum efficiency.
-
