--- keyboards/preonic/keymaps/juno/readme.md
@@ -1,1 +1,0 @@
-# The default Preonic layout - largely based on the Planck's
