--- keyboards/vitamins_included/keymaps/juno/rules.mk
@@ -1,16 +1,0 @@
-MIDI_ENABLE=no
-BACKLIGHT_ENABLE = no
-RGBLIGHT_ENABLE = no
-
-NKRO_ENABLE = no	    # USB Nkey Rollover
-AUDIO_ENABLE = no
-
-## BOOTMAGIC_ENABLE = yes	# Virtual DIP switch configuration
-## (Note that for BOOTMAGIC on Teensy LC you have to use a custom .ld script.)
-
-MOUSEKEY_ENABLE = no	# Mouse keys
-EXTRAKEY_ENABLE = yes	# Audio control and System control
-CONSOLE_ENABLE = no		# Console for debug
-COMMAND_ENABLE = no    	# Commands for debug and configuration
-
-#SLEEP_LED_ENABLE = yes  # Breathing sleep LED during USB suspend
