+++ layouts/community/ortho_4x12/juno/readme.md
@@ -1,0 +1,16 @@
+# Juno's Universal ortholinear 4x12 Layout
+created by Juno Nguyen
+
+## Features:
+* **Minimal changes from the default keymap**: to minimize re-adaptation.
+* **Top corner Esc**: swapped with Tab.
+* **Double function ShiftEnter**: hold for Right Shift; tap for Enter.
+* **Removed infrequently used keys**: Euro-specific keys.
+* **Custom Fn layer**: inspired by Pok3r and refined over time.
+* **SpaceFn**: the ultimate way to Fn.
+* **Double arrow clusters**: both Fn + WASD and Fn + IJKL are simultaneously usable for maximum comfort.
+* **Gaming-specific layer**: T-inverted arrow cluster shape; swapped Raise and Lower for easy access to numeric keys (for switching weapons in action games); standard-positioned bottom modifiers.
+* **LED indicator**: lighting up when Fn or Gaming layer is activated.
+* **Audio indicators**: Caps Lock and gaming layer.
+
+Refined since September 2017.
