+++ layouts/community/ortho_4x12/junonum/config.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #ifdef AUDIO_ENABLE
-    #define STARTUP_SONG SONG(PREONIC_SOUND)
+    #define STARTUP_SONG SONG(PLANCK_SOUND)
     // #define STARTUP_SONG SONG(NO_SOUND)
 
     #define DEFAULT_LAYER_SONGS { SONG(QWERTY_SOUND), \
@@ -9,8 +9,6 @@
                                   SONG(DVORAK_SOUND) \
                                 }
 #endif
-
-#define MUSIC_MASK (keycode != KC_NO)
 
 /*
  * MIDI options
@@ -22,7 +20,6 @@
 /* enable basic MIDI features:
    - MIDI notes can be sent when in Music mode is on
 */
-
 #define MIDI_BASIC
 
 /* enable advanced MIDI features:
@@ -35,3 +32,9 @@
 
 /* override number of MIDI tone keycodes (each octave adds 12 keycodes and allocates 12 bytes) */
 //#define MIDI_TONE_KEYCODE_OCTAVES 2
+
+// Most tactile encoders have detents every 4 stages
+#define ENCODER_RESOLUTION 4
+
+// Reduce power consumption for iOS compatibility
+#define USB_MAX_POWER_CONSUMPTION 100
