+++ layouts/community/ortho_4x12/junonum/readme.md
@@ -1,0 +1,11 @@
+# junonum
+A universal ortholinear 4x12 keymap created by Juno Nguyen
+
+## Features:
+* Successor to *juno*
+* Numpad-like cluster for the numerics.
+* ESDF and IJKL arrow cluster from SpaceFn
+* Huge amount of omissions of the redundant keys.
+* minimalism (not really).
+
+Refined since Oct 2019
